Split DAHD and ADSR
DAHD and ADSR and ADAR share a lot of code but share even more
now while making DAHD and ADSR separate envelope types.
