+++ include/sst/basic-blocks/dsp/PanLaws.h
@@ -15,6 +15,8 @@
  * All source for sst-basic-blocks is available at
  * https://github.com/surge-synthesizer/sst-basic-blocks
  */
+
+
 #ifndef SST_BASIC_BLOCKS_DSP_PANLAWS_H
 #define SST_BASIC_BLOCKS_DSP_PANLAWS_H
