+++ include/sst/basic-blocks/modulators/ADAREnvelope.h
@@ -16,48 +16,40 @@
  * https://github.com/surge-synthesizer/sst-basic-blocks
  */
 
+
+
 #ifndef SST_BASIC_BLOCKS_MODULATORS_ADARENVELOPE_H
 #define SST_BASIC_BLOCKS_MODULATORS_ADARENVELOPE_H
 
 #include <algorithm>
 #include <cmath>
 #include <cassert>
+#include "DiscreteStagesEnvelope.h"
 
 namespace sst::basic_blocks::modulators
 {
-// For context on SRProvider see the ADSRDAHD Envelope
-template<typename SRProvider, int BLOCK_SIZE>
-struct ADAREnvelope
+/**
+ * ADAREnvelope is an attack decay or a gated attack release envelope, the only
+ * difference being a hold-while-gated approach.
+ *
+ * @tparam SRProvider See the comments in ADSREnvelope
+ * @tparam BLOCK_SIZE Must be a power of 2
+ * @tparam RangeProvider Defines the min and max
+ */
+template<typename SRProvider, int BLOCK_SIZE, typename RangeProvider = TenSecondRange>
+struct ADAREnvelope : DiscreteStagesEnvelope<BLOCK_SIZE, RangeProvider>
 {
+    using base_t = DiscreteStagesEnvelope<BLOCK_SIZE, RangeProvider>;
+
     SRProvider *srProvider;
     ADAREnvelope(SRProvider *s) : srProvider(s)
     {
-        for (int i = 0; i < BLOCK_SIZE; ++i)
-            outputCache[i] = 0;
     }
-    enum Stage
-    {
-        s_attack,
-        s_hold,
-        s_decay,
-        s_analog_residual_decay,
-        s_eoc,
-        s_complete
-    } stage{s_complete};
-
-
-    static_assert((BLOCK_SIZE >= 8) & !(BLOCK_SIZE & (BLOCK_SIZE - 1)), "Block size must be power of 2 8 or above.");
-    static constexpr float BLOCK_SIZE_INV{1.f/BLOCK_SIZE};
 
     bool isDigital{true};
     bool isGated{false};
 
     float phase{0}, start{0};
-
-    float output{0}, eoc_output{0};
-    float outputCache[BLOCK_SIZE], outBlock0{0.f};
-    int current{BLOCK_SIZE};
-    int eoc_countdown{0};
 
     // Analog Mode
     float v_c1{0}, v_c1_delayed{0.f};
@@ -82,52 +74,50 @@ struct ADAREnvelope
             }
         }
         phase = f;
-        stage = s_attack;
-        current = BLOCK_SIZE;
-        outBlock0 = f;
+        this->stage = base_t::s_attack;
         isDigital = id;
         isGated = ig;
-        eoc_output = 0;
-        eoc_countdown = 0;
 
         v_c1 = f;
         v_c1_delayed = f;
         discharge = false;
+
+        base_t::resetCurrent();
     }
 
     template <bool gated>
     inline float stepDigital(const float a, const float d, const bool gateActive)
     {
         float target = 0;
-        switch (stage)
+        switch (this->stage)
         {
         default:
             break;
-        case s_attack:
+        case base_t::s_attack:
         {
             phase += srProvider->envelope_rate_linear_nowrap(a);
             if (phase >= 1)
             {
                 phase = 1;
                 if constexpr (gated)
-                    stage = s_hold;
+                    this->stage = base_t::s_hold;
                 else
-                    stage = s_decay;
+                    this->stage = base_t::s_decay;
             }
             if constexpr (gated)
                 if (!gateActive)
-                    stage = s_decay;
+                    this->stage = base_t::s_decay;
             target = phase;
             break;
         }
-        case s_decay:
+        case base_t::s_decay:
         {
             phase -= srProvider->envelope_rate_linear_nowrap(d);
             if (phase <= 0)
             {
                 phase = 0;
-                stage = s_eoc;
-                eoc_countdown = (int)std::round(srProvider->samplerate * 0.01);
+                this->stage = base_t::s_eoc;
+                this->eoc_countdown = (int)std::round(srProvider->samplerate * 0.01);
             }
             target = phase;
         }
@@ -135,55 +125,30 @@ struct ADAREnvelope
         return target;
     }
 
-    inline void immediatelyEnd()
+    inline void process01AD(const float a, const float d, const int ashape, const int dshape,
+                            const bool gateActive)
     {
-        stage = s_complete;
-        eoc_output = 0;
-        output = 0;
-        eoc_countdown = 0;
-        current = BLOCK_SIZE;
-        outBlock0 = 0;
+        processScaledAD(this->rateFrom01(1), this->rateFrom01(d), ashape, dshape, gateActive);
+
     }
-    inline void process(const float a, const float d, const int ashape, const int dshape,
+
+    inline void processScaledAD(const float a, const float d, const int ashape, const int dshape,
                         const bool gateActive)
     {
-        if (stage == s_complete)
-        {
-            output = 0;
+        if (base_t::preBlockCheck())
             return;
-        }
 
-        if (stage == s_eoc)
-        {
-            output = 0;
-            eoc_output = 1;
-
-            eoc_countdown--;
-            if (eoc_countdown == 0)
-            {
-                eoc_output = 0;
-                stage = s_complete;
-            }
-            return;
-        }
-        eoc_output = 0;
-
-        if (stage == s_analog_residual_decay && eoc_countdown)
-        {
-            eoc_output = 1;
-            eoc_countdown--;
-        }
-        if (current == BLOCK_SIZE)
+        if (this->current == BLOCK_SIZE)
         {
             float target = 0;
-            if (isGated && stage == s_hold)
+            if (isGated && this->stage == base_t::s_hold)
             {
                 target = 1;
 
                 if (!gateActive)
                 {
                     phase = 1;
-                    stage = s_decay;
+                    this->stage = base_t::s_decay;
                 }
             }
             else if (isDigital)
@@ -195,13 +160,13 @@ struct ADAREnvelope
             }
             else
             {
-                const float coeff_offset = 2.f - std::log2(srProvider->samplerate * BLOCK_SIZE_INV);
+                const float coeff_offset = 2.f - std::log2(srProvider->samplerate * base_t::BLOCK_SIZE_INV);
 
                 auto ndc = (v_c1_delayed >= 0.99999f);
                 if (ndc && !discharge)
                 {
                     phase = 1;
-                    stage = isGated ? s_hold : s_decay;
+                    this->stage = isGated ? base_t::s_hold : base_t::s_decay;
                 }
 
                 if (isGated && !discharge)
@@ -215,7 +180,7 @@ struct ADAREnvelope
                 auto v_attack = (!discharge) * v_gate;
                 auto v_decay = (!discharge) * v_gate;
 
-                // In this case we only need the coefs in their stage
+                // In this case we only need the coefs in their this->stage
                 float coef_A = !discharge ? powf(2.f, std::min(0.f, coeff_offset - a)) : 0;
                 float coef_D = discharge ? powf(2.f, std::min(0.f, coeff_offset - d)) : 0;
 
@@ -225,13 +190,13 @@ struct ADAREnvelope
                 v_c1 = v_c1 + diff_v_a * coef_A + diff_v_d * coef_D;
                 target = v_c1;
 
-                if (stage == s_decay)
+                if (this->stage == base_t::s_decay)
                 {
                     phase -= srProvider->envelope_rate_linear_nowrap(d);
                     if (phase <= 0)
                     {
-                        eoc_countdown = (int)std::round(srProvider->samplerate * 0.01);
-                        stage = s_analog_residual_decay;
+                        this->eoc_countdown = (int)std::round(srProvider->samplerate * 0.01);
+                        this->stage = base_t::s_analog_residual_decay;
                     }
                 }
                 if (v_c1 < 1e-6 && discharge)
@@ -240,58 +205,28 @@ struct ADAREnvelope
                     v_c1_delayed = 0;
                     discharge = false;
                     target = 0;
-                    if (stage != s_analog_residual_decay)
+                    if (this->stage != base_t::s_analog_residual_decay)
                     {
-                        eoc_countdown = (int)std::round(srProvider->samplerate * 0.01);
-                        stage = s_eoc;
+                        this->eoc_countdown = (int)std::round(srProvider->samplerate * 0.01);
+                        this->stage = base_t::s_eoc;
                     }
                     else
                     {
-                        stage = s_complete;
-                        eoc_countdown = 0;
+                        this->stage = base_t::s_complete;
+                        this->eoc_countdown = 0;
                     }
                 }
             }
 
             if (isDigital)
             {
-                if (stage == s_attack)
-                {
-                    switch (ashape)
-                    {
-                    case 0:
-                        target = sqrt(target);
-                        break;
-                    case 2:
-                        target = target * target * target;
-                        break;
-                    }
-                }
-                else
-                {
-                    switch (dshape)
-                    {
-                    case 0:
-                        target = sqrt(target);
-                        break;
-                    case 2:
-                        target = target * target * target;
-                        break;
-                    }
-                }
+                target = base_t::shapeTarget(target, ashape, dshape);
             }
-            float dO = (target - outBlock0) * BLOCK_SIZE_INV;
-            for (int i = 0; i < BLOCK_SIZE; ++i)
-            {
-                outputCache[i] = outBlock0 + dO * i;
-            }
-            outBlock0 = target;
 
-            current = 0;
+            base_t::updateBlockTo(target);
         }
 
-        output = outputCache[current];
-        current++;
+        base_t::step();
     }
 };
 } // namespace sst::surgext_rack::dsp::envelopes
