+++ include/sst/basic-blocks/modulators/SimpleLFO.h
@@ -16,6 +16,8 @@
  * https://github.com/surge-synthesizer/sst-basic-blocks
  */
 
+
+
 #ifndef SST_BASIC_BLOCKS_MODULATORS_SIMPLELFO_H
 #define SST_BASIC_BLOCKS_MODULATORS_SIMPLELFO_H
